pyproject.toml
[project]
name = "z-iamge-world"
version = "0.1.0"
description = "Interactive world model using Z-Image and StreamDiffusion"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "accelerate>=1.0.0",
    "transformers>=4.45.0",
    "safetensors>=0.4.0",
    "einops>=0.8.0",
    "omegaconf>=2.3.0",
    "pyyaml>=6.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
training = [
    "peft>=0.13.0",
    "wandb>=0.18.0",
]
demo = [
    "pygame>=2.6.0",
]
video = [
    "opencv-python>=4.10.0",
    "imageio>=2.36.0",
    "imageio-ffmpeg>=0.5.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
all = [
    "peft>=0.13.0",
    "wandb>=0.18.0",
    "pygame>=2.6.0",
    "opencv-python>=4.10.0",
    "imageio>=2.36.0",
    "imageio-ffmpeg>=0.5.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["models", "streaming", "training", "inference", "data", "demo"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
